Boot postgres setup
Motivation:

We want to access postgres.

Modifications:

Added necessary dependencies. Wasn't able to use ClanHR's utility lib
because of conflicts with luminus, so defaulted to simple async query.
Also, don't really know how to work with luminus config, so did a
martelate do load the specific database config.

Result:

Added a test that succeedes if the PG connection is available.