Make migratus work
Motivation:

We need to be able to run migrations

Modifications:

The project wasn't added with +postgres support and was lacking the
needed to run migratus. Added the dep and configured new alias: `lein
migrate` and `lein rollback`.

Also added duplicate information on the conn string info because jdbc
need extra information.

And changed travis run script to run migrations before running the
tests.